+++ rows/plugins/xls.py
@@ -1,18 +1,18 @@
 # coding: utf-8
 
-# Copyright 2014-2016 Álvaro Justen <https://github.com/turicas/rows/>
-#
+# Copyright 2014-2018 Álvaro Justen <https://github.com/turicas/rows/>
+
 #    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU General Public License as published by
+#    it under the terms of the GNU Lesser General Public License as published by
 #    the Free Software Foundation, either version 3 of the License, or
 #    (at your option) any later version.
-#
+
 #    This program is distributed in the hope that it will be useful,
 #    but WITHOUT ANY WARRANTY; without even the implied warranty of
 #    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
-#
-#    You should have received a copy of the GNU General Public License
+#    GNU Lesser General Public License for more details.
+
+#    You should have received a copy of the GNU Lesser General Public License
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from __future__ import unicode_literals
@@ -105,6 +105,9 @@ def cell_value(sheet, row, col):
         elif value == 1:
             return True
 
+    elif cell.xf_index is None:
+        return value  # TODO: test
+
     else:
         book = sheet.book
         xf = book.xf_list[cell.xf_index]
@@ -132,8 +135,10 @@ def cell_value(sheet, row, col):
 
 
 def import_from_xls(filename_or_fobj, sheet_name=None, sheet_index=0,
-                    start_row=0, start_column=0, *args, **kwargs):
-    """Return a rows.Table created from imported XLS file."""
+                    start_row=0, start_column=0, end_row=None, end_column=None,
+                    *args, **kwargs):
+      """Return a rows.Table created from imported XLS file."""
+
     filename, _ = get_filename_and_fobj(filename_or_fobj, mode='rb')
     book = xlrd.open_workbook(filename, formatting_info=True)
     if sheet_name is not None:
@@ -143,9 +148,14 @@ def import_from_xls(filename_or_fobj, sheet_name=None, sheet_index=0,
     # TODO: may re-use Excel data types
 
     # Get header and rows
+    # xlrd library reads rows and columns starting from 0 and ending on
+    # sheet.nrows/ncols - 1. rows accepts the same pattern
+    max_row, max_col = sheet.nrows - 1, sheet.ncols - 1
+    column_range = range(start_column, (end_column or max_col) + 1)
+    row_range = range(start_row, (end_row or max_row) + 1)
     table_rows = [[cell_value(sheet, row_index, column_index)
-                   for column_index in range(start_column, sheet.ncols)]
-                  for row_index in range(start_row, sheet.nrows)]
+                   for column_index in column_range]
+                  for row_index in row_range]
 
     meta = {'imported_from': 'xls',
             'filename': filename,