+++ rows/plugins/xlsx.py
@@ -1,18 +1,18 @@
 # coding: utf-8
 
-# Copyright 2014-2016 Álvaro Justen <https://github.com/turicas/rows/>
-#
+# Copyright 2014-2018 Álvaro Justen <https://github.com/turicas/rows/>
+
 #    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU General Public License as published by
+#    it under the terms of the GNU Lesser General Public License as published by
 #    the Free Software Foundation, either version 3 of the License, or
 #    (at your option) any later version.
-#
+
 #    This program is distributed in the hope that it will be useful,
 #    but WITHOUT ANY WARRANTY; without even the implied warranty of
 #    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
-#
-#    You should have received a copy of the GNU General Public License
+#    GNU Lesser General Public License for more details.
+
+#    You should have received a copy of the GNU Lesser General Public License
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from __future__ import unicode_literals
@@ -60,18 +60,27 @@ def _cell_to_python(cell):
 
 
 def import_from_xlsx(filename_or_fobj, sheet_name=None, sheet_index=0,
-                     start_row=0, start_column=0, *args, **kwargs):
+                     start_row=0, start_column=0, end_row=None,
+                     end_column=None, *args, **kwargs):
     """Return a rows.Table created from imported XLSX file."""
-    workbook = load_workbook(filename_or_fobj)
+    
+    workbook = load_workbook(filename_or_fobj, data_only=True)
     if sheet_name is None:
         sheet_name = workbook.sheetnames[sheet_index]
-    sheet = workbook.get_sheet_by_name(sheet_name)
+    sheet = workbook[sheet_name]
 
-    start_row, end_row = max(start_row, sheet.min_row), sheet.max_row
-    start_col, end_col = max(start_column, sheet.min_column), sheet.max_column
+    # openpyxl library reads rows and columns starting from 1 and ending on
+    # sheet.max_row/max_col. rows uses another pattern: 0 to N - 1, so we need
+    # to adjust the ranges accordingly
+    min_row, min_col = sheet.min_row - 1, sheet.min_column - 1
+    max_row, max_col = sheet.max_row - 1, sheet.max_column - 1
+    start_row = max(start_row, min_row)
+    end_row = min(end_row or max_row, max_row)
+    start_col = max(start_column, min_col)
+    end_col = min(end_column or max_col, max_col)
     table_rows = [[_cell_to_python(sheet.cell(row=row_index, column=col_index))
-                   for col_index in range(start_col, end_col + 1)]
-                  for row_index in range(start_row, end_row + 1)]
+                   for col_index in range(start_col + 1, end_col + 2)]
+                  for row_index in range(start_row + 1, end_row + 2)]
 
     filename, _ = get_filename_and_fobj(filename_or_fobj, dont_open=True)
     metadata = {'imported_from': 'xlsx',
@@ -118,6 +127,7 @@ def _python_to_cell(field_types):
 def export_to_xlsx(table, filename_or_fobj=None, sheet_name='Sheet1', *args,
                    **kwargs):
     """Export the rows.Table to XLSX file and return the saved file."""
+    
     workbook = Workbook()
     sheet = workbook.active
     sheet.title = sheet_name