+++ tests/tests_plugin_csv.py
@@ -1,18 +1,18 @@
 # coding: utf-8
 
 # Copyright 2014-2017 Álvaro Justen <https://github.com/turicas/rows/>
-#
+
 #    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU General Public License as published by
+#    it under the terms of the GNU Lesser General Public License as published by
 #    the Free Software Foundation, either version 3 of the License, or
 #    (at your option) any later version.
-#
+
 #    This program is distributed in the hope that it will be useful,
 #    but WITHOUT ANY WARRANTY; without even the implied warranty of
 #    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
-#
-#    You should have received a copy of the GNU General Public License
+#    GNU Lesser General Public License for more details.
+
+#    You should have received a copy of the GNU Lesser General Public License
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from __future__ import unicode_literals
@@ -52,8 +52,10 @@ class PluginCsvTestCase(utils.RowsTestMixIn, unittest.TestCase):
     assert_meta_encoding = True
 
     def test_imports(self):
-        self.assertIs(rows.import_from_csv, rows.plugins.plugin_csv.import_from_csv)
-        self.assertIs(rows.export_to_csv, rows.plugins.plugin_csv.export_to_csv)
+        self.assertIs(rows.import_from_csv,
+                      rows.plugins.plugin_csv.import_from_csv)
+        self.assertIs(rows.export_to_csv,
+                      rows.plugins.plugin_csv.export_to_csv)
 
     @mock.patch('rows.plugins.plugin_csv.create_table')
     def test_import_from_csv_uses_create_table(self, mocked_create_table):
@@ -68,7 +70,7 @@ class PluginCsvTestCase(utils.RowsTestMixIn, unittest.TestCase):
         call = mocked_create_table.call_args
         kwargs['meta'] = {'imported_from': 'csv',
                           'filename': self.filename,
-                          'encoding': 'utf-8',}
+                          'encoding': 'utf-8'}
         self.assertEqual(call[1], kwargs)
 
     @mock.patch('rows.plugins.plugin_csv.create_table')
@@ -112,7 +114,7 @@ class PluginCsvTestCase(utils.RowsTestMixIn, unittest.TestCase):
 
         # Should not raise `UnicodeDecodeError`
         table = rows.import_from_csv(BytesIO(data), encoding='utf-8',
-                sample_size=262144)
+                                     sample_size=262144)
 
         last_row = table[-1]
         last_column = 'b' * 508
@@ -312,3 +314,14 @@ class PluginCsvTestCase(utils.RowsTestMixIn, unittest.TestCase):
         result_1 = rows.export_to_csv(utils.table, dialect=csv.excel_tab)
         result_2 = rows.export_to_csv(utils.table, dialect=csv.excel)
         self.assertEqual(result_1.replace(b'\t', b','), result_2)
+
+    def test_export_callback(self):
+        table = rows.import_from_dicts([{'id': number}
+                                        for number in range(10)])
+        myfunc = mock.Mock()
+        rows.export_to_csv(table, callback=myfunc, batch_size=3)
+        self.assertEqual(myfunc.call_count, 4)
+        self.assertEqual(
+            [x[0][0] for x in myfunc.call_args_list],
+            [3, 6, 9, 10]
+        )