+++ tests/tests_plugin_sqlite.py
@@ -1,18 +1,18 @@
 # coding: utf-8
 
-# Copyright 2014-2016 Álvaro Justen <https://github.com/turicas/rows/>
-#
+# Copyright 2014-2018 Álvaro Justen <https://github.com/turicas/rows/>
+
 #    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU General Public License as published by
+#    it under the terms of the GNU Lesser General Public License as published by
 #    the Free Software Foundation, either version 3 of the License, or
 #    (at your option) any later version.
-#
+
 #    This program is distributed in the hope that it will be useful,
 #    but WITHOUT ANY WARRANTY; without even the implied warranty of
 #    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
-#
-#    You should have received a copy of the GNU General Public License
+#    GNU Lesser General Public License for more details.
+
+#    You should have received a copy of the GNU Lesser General Public License
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from __future__ import unicode_literals
@@ -194,3 +194,14 @@ class PluginSqliteTestCase(utils.RowsTestMixIn, unittest.TestCase):
                 query_args=(3, ))
         for row in table:
             self.assertTrue(row.float_column > 3)
+
+    def test_export_callback(self):
+        table = rows.import_from_dicts([{'id': number}
+                                        for number in range(10)])
+        myfunc = mock.Mock()
+        rows.export_to_sqlite(table, ':memory:', callback=myfunc, batch_size=3)
+        self.assertEqual(myfunc.call_count, 4)
+        self.assertEqual(
+            [x[0][0] for x in myfunc.call_args_list],
+            [3, 6, 9, 10]
+        )