+++ tests/tests_plugin_utils.py
@@ -1,18 +1,18 @@
 # coding: utf-8
 
-# Copyright 2014-2015 Álvaro Justen <https://github.com/turicas/rows/>
-#
+# Copyright 2014-2017 Álvaro Justen <https://github.com/turicas/rows/>
+
 #    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU General Public License as published by
+#    it under the terms of the GNU Lesser General Public License as published by
 #    the Free Software Foundation, either version 3 of the License, or
 #    (at your option) any later version.
-#
+
 #    This program is distributed in the hope that it will be useful,
 #    but WITHOUT ANY WARRANTY; without even the implied warranty of
 #    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
-#
-#    You should have received a copy of the GNU General Public License
+#    GNU Lesser General Public License for more details.
+
+#    You should have received a copy of the GNU Lesser General Public License
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from __future__ import unicode_literals
@@ -36,6 +36,7 @@ from rows import fields
 class GenericUtilsTestCase(unittest.TestCase):
 
     def test_slug(self):
+        self.assertEqual(plugins_utils.slug(None), '')
         self.assertEqual(plugins_utils.slug('Álvaro Justen'), 'alvaro_justen')
         self.assertEqual(plugins_utils.slug("Moe's Bar"), 'moes_bar')
         self.assertEqual(plugins_utils.slug("-----te-----st------"), 'te_st')
@@ -45,6 +46,7 @@ class GenericUtilsTestCase(unittest.TestCase):
                 'query_occurrence_first_seen')
         self.assertEqual(plugins_utils.slug(' ÁLVARO  justen% '),
                          'alvaro_justen')
+        self.assertEqual(plugins_utils.slug(42), '42')
 
     def test_ipartition(self):
         iterable = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]