+++ tests/utils.py
@@ -1,18 +1,18 @@
 # coding: utf-8
 
-# Copyright 2014-2016 Álvaro Justen <https://github.com/turicas/rows/>
-#
+# Copyright 2014-2017 Álvaro Justen <https://github.com/turicas/rows/>
+
 #    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU General Public License as published by
+#    it under the terms of the GNU Lesser General Public License as published by
 #    the Free Software Foundation, either version 3 of the License, or
 #    (at your option) any later version.
-#
+
 #    This program is distributed in the hope that it will be useful,
 #    but WITHOUT ANY WARRANTY; without even the implied warranty of
 #    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
-#
-#    You should have received a copy of the GNU General Public License
+#    GNU Lesser General Public License for more details.
+
+#    You should have received a copy of the GNU Lesser General Public License
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from __future__ import unicode_literals
@@ -177,6 +177,12 @@ class RowsTestMixIn(object):
                              'filename': filename,}
             if self.assert_meta_encoding:
                 expected_meta['encoding'] = self.encoding
+
+        # Don't test 'frame_style' metadata,
+        # as it is specific for txt importing
+        # (and the default values for it might change)
+        if "frame_style" not in expected_meta:
+            kwargs['meta'].pop('frame_style', '')
 
         self.assertEqual(kwargs['meta'], expected_meta)
         del kwargs['meta']